+++ packages/salesforcedx-vscode-core/src/diagnostics/diagnostics.ts
@@ -4,7 +4,7 @@
  * Licensed under the BSD 3-Clause license.
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-import { ExecuteAnonymousResponse } from '@salesforce/salesforcedx-apex/packages/apex';
+import { ExecuteAnonymousResponse } from '@salesforce/apex-node';
 import { ForceSourceDeployErrorResponse } from '@salesforce/salesforcedx-utils-vscode/out/src/cli';
 import { DeployResult } from '@salesforce/source-deploy-retrieve';
 import * as path from 'path';
